--- src/feature/auth/dto/sendCodeResponse.dto.ts
@@ -1,7 +1,0 @@
-export class SendCodeResponseDto {
-    email: string;
-    state: {
-        code: string,
-        verified: boolean
-    };
-}
